refactor(chat): extract dialogs and sheets from ChatScreen
Move the preset picker/save, fork options, model parameters, rename/delete
confirmations, and the primary + secondary model selector sheets into
ChatScreenDialogs, along with the ChatRenameDialog/ChatDeleteConfirmationDialog
composables. ChatScreen passes the open-state flags and their setters.
