+++ feature/chat/src/androidMain/kotlin/com/garfiec/librechat/feature/chat/screen/ChatScreen.kt
@@ -10,7 +10,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -32,7 +31,6 @@ import androidx.compose.material.icons.outlined.FileOpen
 import androidx.compose.material.icons.outlined.PhotoLibrary
 import androidx.compose.material.icons.outlined.SaveAs
 import androidx.compose.material.icons.outlined.Share
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -40,12 +38,10 @@ import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -59,17 +55,11 @@ import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.garfiec.librechat.core.common.EndpointConstants
 import com.garfiec.librechat.core.data.datastore.ChatFontSize
 import com.garfiec.librechat.core.data.datastore.LatexRenderer
-import com.garfiec.librechat.core.ui.components.ModelParameterSheet
 import com.garfiec.librechat.feature.chat.components.ChatInput
 import com.garfiec.librechat.feature.chat.components.ChatRoot
-import com.garfiec.librechat.feature.chat.components.ForkOptionsBottomSheet
 import com.garfiec.librechat.feature.chat.components.InConvoSearchBar
-import com.garfiec.librechat.feature.chat.components.ModelSelectorSheet
-import com.garfiec.librechat.feature.chat.components.PresetPicker
-import com.garfiec.librechat.feature.chat.components.SavePresetDialog
 import com.garfiec.librechat.feature.chat.components.TempChatToggle
 import com.garfiec.librechat.feature.chat.resources.*
 import com.garfiec.librechat.feature.chat.resources.Res
@@ -291,150 +281,18 @@ actual fun ChatScreen(
         }
     }
 
-    if (showPresetPicker) {
-        PresetPicker(
-            presets = uiState.presets,
-            onPresetSelect = { preset ->
-                viewModel.loadPreset(preset)
-                showPresetPicker = false
-            },
-            onDismiss = { showPresetPicker = false },
-            onEditPreset = { preset ->
-                viewModel.loadPreset(preset)
-                showPresetPicker = false
-            },
-            onDeletePreset = { preset ->
-                preset.presetId?.let { viewModel.deletePreset(it) }
-            },
-        )
-    }
-
-    if (showSavePresetDialog) {
-        SavePresetDialog(
-            currentEndpoint = uiState.selectedEndpoint,
-            currentModel = uiState.selectedModel,
-            onSave = { name ->
-                viewModel.savePreset(name)
-                showSavePresetDialog = false
-            },
-            onDismiss = { showSavePresetDialog = false },
-        )
-    }
-
-    if (uiState.showForkOptionsForMessageId != null) {
-        ForkOptionsBottomSheet(
-            onDismiss = viewModel::dismissForkOptions,
-            onFork = { option, splitAtTarget ->
-                viewModel.forkFromMessage(
-                    messageId = uiState.showForkOptionsForMessageId!!,
-                    option = option,
-                    splitAtTarget = splitAtTarget,
-                )
-            },
-        )
-    }
-
-    if (uiState.showModelParameters) {
-        val activeAgent = remember(uiState.agents, uiState.selectedModel, uiState.selectedEndpoint) {
-            if (uiState.selectedEndpoint == EndpointConstants.AGENTS) {
-                uiState.agents.find { it.id == uiState.selectedModel }
-            } else {
-                null
-            }
-        }
-        ModelParameterSheet(
-            parameters = uiState.modelParameters,
-            onParametersChange = viewModel::updateModelParameters,
-            onDismiss = viewModel::hideModelParameters,
-            selectedEndpoint = uiState.selectedEndpoint,
-            extendedEffortSupported = uiState.extendedEffortSupported,
-            selectedProvider = activeAgent?.provider,
-            selectedModel = activeAgent?.model ?: uiState.selectedModel,
-            onSaveAsPreset = {
-                viewModel.hideModelParameters()
-                showSavePresetDialog = true
-            },
-        )
-    }
-
-    if (uiState.showRenameDialog) {
-        ChatRenameDialog(
-            currentTitle = uiState.conversationTitle ?: "",
-            onDismiss = viewModel::dismissRenameDialog,
-            onConfirm = viewModel::renameConversation,
-        )
-    }
-
-    if (uiState.showDeleteConfirmation) {
-        ChatDeleteConfirmationDialog(
-            conversationTitle = uiState.conversationTitle ?: "this conversation",
-            onDismiss = viewModel::dismissDeleteConfirmation,
-            onConfirm = viewModel::deleteConversation,
-        )
-    }
-
-    if (uiState.showModelSheet) {
-        ModelSelectorSheet(
-            endpointConfigs = uiState.endpointConfigs,
-            availableModels = uiState.availableModels,
-            agents = uiState.agents,
-            selectedEndpoint = uiState.selectedEndpoint,
-            selectedModel = uiState.selectedModel,
-            onModelSelect = { endpoint, model ->
-                viewModel.onModelSelected(endpoint, model)
-                // Clear any pending scaffold-level snackbar for the same error so it
-                // doesn't flash behind the sheet's close animation. Harmless no-op when
-                // error is already null.
-                viewModel.dismissError()
-                viewModel.dismissSendBlockReason()
-                viewModel.dismissModelSheet()
-            },
-            onDismiss = {
-                viewModel.dismissError()
-                viewModel.dismissSendBlockReason()
-                viewModel.dismissModelSheet()
-            },
-            serverUrl = uiState.serverUrl,
-            // Send-block reasons take precedence: when set, the sheet was auto-opened
-            // to help the user resolve the block, so surface that context inline.
-            errorMessage = sendBlockMessage ?: uiState.error,
-            onErrorDismiss = {
-                viewModel.dismissSendBlockReason()
-                viewModel.dismissError()
-            },
-            favoriteAgentIds = uiState.favoriteAgentIds,
-            favoriteModelKeys = uiState.favoriteModelKeys,
-            onToggleAgentFavorite = viewModel::toggleAgentFavorite,
-            onToggleModelFavorite = viewModel::toggleModelFavorite,
-            starredDisplay = uiState.starredModelsDisplay,
-            endpointKeyStates = uiState.endpointKeyStates,
-            onSetApiKey = { name -> onNavigateToProviderKeys(name) },
-        )
-    }
-
-    // Secondary model selector sheet for comparison mode
-    if (showSecondaryModelSheet) {
-        ModelSelectorSheet(
-            endpointConfigs = uiState.endpointConfigs,
-            availableModels = uiState.availableModels,
-            agents = uiState.agents,
-            selectedEndpoint = uiState.comparisonState.secondaryEndpoint,
-            selectedModel = uiState.comparisonState.secondaryModel,
-            onModelSelect = { endpoint, model ->
-                viewModel.setSecondaryModel(endpoint, model)
-                showSecondaryModelSheet = false
-            },
-            onDismiss = { showSecondaryModelSheet = false },
-            serverUrl = uiState.serverUrl,
-            favoriteAgentIds = uiState.favoriteAgentIds,
-            favoriteModelKeys = uiState.favoriteModelKeys,
-            onToggleAgentFavorite = viewModel::toggleAgentFavorite,
-            onToggleModelFavorite = viewModel::toggleModelFavorite,
-            starredDisplay = uiState.starredModelsDisplay,
-            endpointKeyStates = uiState.endpointKeyStates,
-            onSetApiKey = { name -> onNavigateToProviderKeys(name) },
-        )
-    }
+    ChatScreenDialogs(
+        uiState = uiState,
+        viewModel = viewModel,
+        sendBlockMessage = sendBlockMessage,
+        showPresetPicker = showPresetPicker,
+        showSavePresetDialog = showSavePresetDialog,
+        showSecondaryModelSheet = showSecondaryModelSheet,
+        onSetShowPresetPicker = { showPresetPicker = it },
+        onSetShowSavePresetDialog = { showSavePresetDialog = it },
+        onSetShowSecondaryModelSheet = { showSecondaryModelSheet = it },
+        onNavigateToProviderKeys = onNavigateToProviderKeys,
+    )
     }
 }
 
@@ -685,79 +543,4 @@ private fun ChatTopBar(
             }
         }
     }
-}
-
-@Composable
-private fun ChatRenameDialog(
-    currentTitle: String,
-    onDismiss: () -> Unit,
-    onConfirm: (String) -> Unit,
-) {
-    var title by remember { mutableStateOf(currentTitle) }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text(stringResource(Res.string.dialog_title_rename)) },
-        text = {
-            Column {
-                Text(
-                    text = "Enter a new title for this conversation.",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                OutlinedTextField(
-                    value = title,
-                    onValueChange = { title = it },
-                    singleLine = true,
-                    modifier = Modifier.fillMaxWidth(),
-                    label = { Text(stringResource(Res.string.hint_title)) },
-                )
-            }
-        },
-        confirmButton = {
-            TextButton(
-                onClick = { onConfirm(title) },
-                enabled = title.isNotBlank(),
-            ) {
-                Text(stringResource(Res.string.action_rename))
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text(stringResource(Res.string.cancel))
-            }
-        },
-    )
-}
-
-@Composable
-private fun ChatDeleteConfirmationDialog(
-    conversationTitle: String,
-    onDismiss: () -> Unit,
-    onConfirm: () -> Unit,
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text(stringResource(Res.string.dialog_title_delete_conversation)) },
-        text = {
-            Text(
-                text = "Are you sure you want to delete \"$conversationTitle\"? This action cannot be undone.",
-                style = MaterialTheme.typography.bodyMedium,
-            )
-        },
-        confirmButton = {
-            TextButton(onClick = onConfirm) {
-                Text(
-                    text = stringResource(Res.string.delete),
-                    color = MaterialTheme.colorScheme.error,
-                )
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text(stringResource(Res.string.cancel))
-            }
-        },
-    )
 }
